requirements.txt
# Copyright 2017 AT&T Intellectual Property.  All other rights reserved.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

alembic==0.9.5
arrow==0.10.0
configparser==3.5.0
falcon==1.2.0
jsonschema==2.6.0
keystoneauth1==2.13.0
keystonemiddleware==4.17.0
oslo.config==4.11.0
oslo.policy==1.25.1
PasteDeploy==1.5.2
pbr!=2.1.0,>=2.0.0 # Apache-2.0
psycopg2==2.7.3.1
python-dateutil==2.6.1
python-memcached==1.58
python-openstackclient==3.11.0
requests==2.18.4
SQLAlchemy==1.1.13
ulid==1.1
uwsgi==2.0.15
click==6.7
click-default-group==1.2
pyyaml==3.12
